requirements.txt
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
bcrypt==3.1.7
black==19.10b0
bson==0.5.9
cassandra-driver==3.22.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Selfdoc==1.2.3
geomet==0.1.2
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
numpy==1.18.4
opencv-python==4.2.0.34
packaging==20.3
pandas==1.0.3
pathspec==0.8.0
pluggy==0.13.1
py==1.8.1
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.7
pytest==5.4.1
pytest-asyncio==0.10.0
python-dateutil==2.8.1
pytz==2020.1
regex==2020.4.4
requests==2.23.0
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.8
uWSGI==2.0.18
wcwidth==0.1.9
Werkzeug==1.0.0
wrapt==1.11.2
zipp==3.1.0
